+++ database/factories/CompanyFactory.php
@@ -7,7 +7,7 @@ $factory->define(App\Company::class, function (Faker $faker) {
 
 	return [
 		'name' => $faker->unique()->company,
-		'shortened' => strtoupper($faker->unique()->firstname),
+		'shortened' => mb_strtoupper($faker->unique()->firstname),
 		'tax_number' => $faker->unique()->randomNumber(8, true) * pow(10, 6),
 		'document_id' => App\Document::first()->id,
 	];