+++ database/factories/ConsultantPositionFactory.php
@@ -7,7 +7,7 @@ $factory->define(App\ConsultantPosition::class, function (Faker $faker) {
   $position_groups = App\PositionGroup::get()->all();
 
   return [
-    'name' => strtoupper($faker->unique()->catchPhrase),
+    'name' => mb_strtoupper($faker->unique()->catchPhrase),
     'charge_id' => $charges[array_rand($charges)]->id,
     'position_group_id' => $position_groups[array_rand($position_groups)]->id
   ];