+++ resources/assets/js/components/departure/DepartureReport.vue
@@ -146,8 +146,6 @@ export default {
         type: null,
         start_date: this.$moment().format("YYYY-MM-DD"),
         end_date: this.$moment().format("YYYY-MM-DD")
-        // start_date: null,
-        // end_date: null,
       },
       menuDateStart: null,
       dateStart: this.$moment().format("DD/MM/YYYY"),
@@ -163,7 +161,7 @@ export default {
     };
   },
   created() {
-    if (this.$store.getters.currentUser.roles[0].name == "juridica") {
+    if (this.$store.getters.role == "juridica") {
       this.juridica = 1;
     }
   },
@@ -216,7 +214,7 @@ export default {
       }
     },
     async checkUsed () {      
-      let departure_used = await axios.get('/departure/get_departures_used/' + this.$store.getters.currentUser.employee_id);
+      let departure_used = await axios.get('/departure/get_departures_used/' + this.$store.getters.id);
 
       var h1 = this.selectedItem.departure_time.split(":");
       var h2 = this.selectedItem.return_time.split(":");