+++ resources/assets/js/components/user/UserLdap.vue
@@ -109,8 +109,7 @@ export default {
           employee_id: id
         });
         this.$router.go({
-          name: "userIndex",
-          params: this.$store.getters.menuLeft.find(obj => { return obj.href == 'userIndex' }).params
+          name: "userIndex"
         });
       } catch (e) {
         console.log(e);