+++ resources/assets/js/store.js
@@ -1,46 +1,39 @@
-import menu from "./menu.js";
 import moment from 'moment';
 
 export default {
   state: {
-    currentUser: localStorage.getItem('user') || null,    
+    id: localStorage.getItem('id') || null,
+    user: localStorage.getItem('user') || null,
+    role: localStorage.getItem('role') || null,
+    permissions: localStorage.getItem('permissions') || null,
     ldapAuth: JSON.parse(process.env.MIX_LDAP_AUTHENTICATION),
     dateNow: moment().format('Y-MM-DD'),
     token: {
       type: localStorage.getItem('token_type') || null,
       value: localStorage.getItem('token') || null
-    },
-    options: localStorage.getItem('options'),
+    }
   },
   getters: {
     ldapAuth(state) {
       return state.ldapAuth
     },
-    currentUser(state) {
-      return JSON.parse(state.currentUser)
+    id(state) {
+      return JSON.parse(state.id)
     },
-    menuLeft(state) {
-      if (state.currentUser) {
-        if (JSON.parse(state.currentUser).roles[0]) {
-          return menu[JSON.parse(state.currentUser).roles[0].name];
-        } else {
-          return menu['empleado'];
-        }
-        return
-      }
-      return null
+    user(state) {
+      return JSON.parse(state.user)
+    },
+    role(state) {
+      return JSON.parse(state.role)
+    },
+    permissions(state) {
+      return JSON.parse(state.permissions)
     },
     dateNow(state) {
       return state.dateNow
     },
     token(state) {
       return state.token
-    },
-    options(state) {
-      return state.options
-    },
-    role(state) {
-      return JSON.parse(state.currentUser).roles[0].name
     }
   },
   mutations: {
@@ -48,15 +41,22 @@ export default {
       localStorage.removeItem('user')
       localStorage.removeItem('token')
       localStorage.removeItem('token_type')
-      state.currentUser = null
-      state.menuLeft = null
-      state.options = null
+      state.id = null
+      state.user = null
+      state.role = null
+      state.permissions = null
     },
     'login': function (state, data) {
       localStorage.setItem("token", data.token);
       localStorage.setItem("token_type", data.token_type);
+      localStorage.setItem("id", JSON.stringify(data.id));
       localStorage.setItem("user", JSON.stringify(data.user));
-      state.currentUser = localStorage.getItem('user');
+      localStorage.setItem("role", JSON.stringify(data.role));
+      localStorage.setItem("permissions", JSON.stringify(data.permissions));
+      state.user = localStorage.getItem('id');
+      state.user = localStorage.getItem('user');
+      state.role = localStorage.getItem('role');
+      state.permissions = localStorage.getItem('permissions');
       state.token = {
         type: localStorage.getItem('token_type'),
         value: localStorage.getItem('token')
@@ -65,10 +65,6 @@ export default {
     },
     'setDate': function(state, newValue) {
       state.dateNow = newValue;
-    },
-    'setOptions': function(state, data) {
-      localStorage.setItem("options", JSON.stringify(data));
-      state.options = localStorage.getItem('options');
     }
   },
   actions: {